refactor(lsp): split oversized service modules (#3980)
