+++ crates/vize_maestro/src/ide/definition/tests.rs
@@ -1,0 +1,16 @@
+mod basics;
+mod contexts;
+mod template;
+
+use tower_lsp::lsp_types::{GotoDefinitionResponse, Location};
+
+fn scalar_location(response: GotoDefinitionResponse) -> Location {
+    match response {
+        GotoDefinitionResponse::Scalar(location) => location,
+        GotoDefinitionResponse::Array(mut locations) => {
+            assert_eq!(locations.len(), 1);
+            locations.remove(0)
+        }
+        GotoDefinitionResponse::Link(_) => panic!("expected location result"),
+    }
+}
