+++ crates/vize_maestro/src/ide/file_rename/manual.rs
@@ -1,28 +1,28 @@
 //! Manual import-path rewriting for file renames.
 #![allow(clippy::disallowed_types, clippy::disallowed_methods)]
 
+mod path;
+mod script;
+
 use std::{
     collections::{HashMap, HashSet},
     fs,
-    path::{Component, Path, PathBuf},
+    path::{Path, PathBuf},
     sync::Mutex,
 };
 
 use ignore::{WalkBuilder, WalkState};
-use oxc_allocator::Allocator;
-use oxc_ast::ast::{CallExpression, Expression, ImportExpression, Statement, TSImportType};
-use oxc_ast_visit::{Visit, walk};
-use oxc_parser::Parser;
-use oxc_span::SourceType;
 use tower_lsp::lsp_types::{FileRename, Range, TextEdit, Url, WorkspaceEdit};
 
+pub(super) use self::path::{
+    RESOLVABLE_SCRIPT_EXTENSIONS, apply_all_path_renames, candidate_exists, normalize_path_buf,
+    relative_module_path, rewrite_relative_specifier, split_specifier_suffix, strip_extension,
+};
+use self::script::{collect_script_file_edits, collect_vue_edits};
 use crate::{ide::offset_to_position, server::ServerState};
 
 const SCRIPT_EXTENSIONS: &[&str] = &["ts", "tsx", "js", "jsx", "mts", "cts", "mjs", "cjs"];
-pub(super) const RESOLVABLE_SCRIPT_EXTENSIONS: &[&str] = &[
-    "ts", "tsx", "d.ts", "d.mts", "d.cts", "js", "jsx", "mts", "cts", "mjs", "cjs", "vue",
-];
-#[derive(Clone)]
+
 pub(super) struct RenameTarget {
     old_path: PathBuf,
     new_path: PathBuf,
@@ -34,128 +34,12 @@ enum ImporterKind {
     Script,
 }
 
-#[derive(Clone, Copy)]
-enum RenderStyle {
-    Explicit,
-    Extensionless,
-    DirectoryIndex,
-}
-
-#[derive(Clone)]
-struct SpecifierCandidate {
-    resolved: PathBuf,
-    style: RenderStyle,
-}
-
-#[derive(Clone)]
-struct SpecifierOccurrence {
-    start: usize,
-    end: usize,
-    specifier: std::string::String,
-}
-
 struct OpenDocumentSnapshot {
     uri: Url,
     source: std::string::String,
 }
 
 type OpenDocumentSnapshots = HashMap<PathBuf, OpenDocumentSnapshot>;
-
-struct ScriptEditContext<'a> {
-    state: &'a ServerState,
-    current_path: &'a Path,
-    future_path: &'a Path,
-    full_source: &'a str,
-    rename_targets: &'a [RenameTarget],
-}
-
-#[derive(Default)]
-struct ModuleSpecifierCollector {
-    specifiers: Vec<SpecifierOccurrence>,
-}
-
-impl ModuleSpecifierCollector {
-    fn push(&mut self, start: u32, end: u32, specifier: &str) {
-        self.specifiers.push(SpecifierOccurrence {
-            start: start as usize,
-            end: end as usize,
-            specifier: specifier.to_string(),
-        });
-    }
-}
-
-impl<'a> Visit<'a> for ModuleSpecifierCollector {
-    fn visit_program(&mut self, program: &oxc_ast::ast::Program<'a>) {
-        for statement in &program.body {
-            match statement {
-                Statement::ImportDeclaration(decl) => {
-                    self.push(
-                        decl.source.span.start + 1,
-                        decl.source.span.end - 1,
-                        decl.source.value.as_str(),
-                    );
-                }
-                Statement::ExportNamedDeclaration(decl) => {
-                    if let Some(source) = &decl.source {
-                        self.push(
-                            source.span.start + 1,
-                            source.span.end - 1,
-                            source.value.as_str(),
-                        );
-                    }
-                }
-                Statement::ExportAllDeclaration(decl) => {
-                    self.push(
-                        decl.source.span.start + 1,
-                        decl.source.span.end - 1,
-                        decl.source.value.as_str(),
-                    );
-                }
-                _ => {}
-            }
-        }
-
-        walk::walk_program(self, program);
-    }
-
-    fn visit_import_expression(&mut self, expression: &ImportExpression<'a>) {
-        if let Expression::StringLiteral(literal) = &expression.source {
-            self.push(
-                literal.span.start + 1,
-                literal.span.end - 1,
-                literal.value.as_str(),
-            );
-        }
-
-        walk::walk_import_expression(self, expression);
-    }
-
-    fn visit_call_expression(&mut self, expression: &CallExpression<'a>) {
-        if let Expression::Identifier(identifier) = &expression.callee
-            && identifier.name.as_str() == "require"
-            && let Some(oxc_ast::ast::Argument::StringLiteral(literal)) =
-                expression.arguments.first()
-        {
-            self.push(
-                literal.span.start + 1,
-                literal.span.end - 1,
-                literal.value.as_str(),
-            );
-        }
-
-        walk::walk_call_expression(self, expression);
-    }
-
-    fn visit_ts_import_type(&mut self, import_type: &TSImportType<'a>) {
-        self.push(
-            import_type.source.span.start + 1,
-            import_type.source.span.end - 1,
-            import_type.source.value.as_str(),
-        );
-
-        walk::walk_ts_import_type(self, import_type);
-    }
-}
 
 pub(super) fn collect_import_rename_edits(
     state: &ServerState,
@@ -335,270 +219,6 @@ fn process_source(
     Some((uri, edits))
 }
 
-fn collect_vue_edits(
-    state: &ServerState,
-    path: &Path,
-    future_path: &Path,
-    source: &str,
-    rename_targets: &[RenameTarget],
-) -> Vec<TextEdit> {
-    let options = vize_atelier_sfc::SfcParseOptions {
-        filename: path.to_string_lossy().to_string().into(),
-        ..Default::default()
-    };
-
-    let Ok(descriptor) = vize_atelier_sfc::parse_sfc(source, options) else {
-        return Vec::new();
-    };
-
-    let mut edits = Vec::new();
-    let edit_context = ScriptEditContext {
-        state,
-        current_path: path,
-        future_path,
-        full_source: source,
-        rename_targets,
-    };
-
-    if let Some(script) = descriptor.script.as_ref() {
-        edits.extend(collect_script_content_edits(
-            &edit_context,
-            script.content.as_ref(),
-            script_source_type(script.lang.as_deref()),
-            script.loc.start,
-        ));
-    }
-
-    if let Some(script_setup) = descriptor.script_setup.as_ref() {
-        edits.extend(collect_script_content_edits(
-            &edit_context,
-            script_setup.content.as_ref(),
-            script_source_type(script_setup.lang.as_deref()),
-            script_setup.loc.start,
-        ));
-    }
-    edits.extend(super::resources::collect_vue_resource_edits(
-        state,
-        path,
-        future_path,
-        source,
-        rename_targets,
-    ));
-
-    edits
-}
-
-fn collect_script_file_edits(
-    state: &ServerState,
-    path: &Path,
-    future_path: &Path,
-    source: &str,
-    rename_targets: &[RenameTarget],
-) -> Vec<TextEdit> {
-    let Some(source_type) = SourceType::from_path(path).ok() else {
-        return Vec::new();
-    };
-
-    let edit_context = ScriptEditContext {
-        state,
-        current_path: path,
-        future_path,
-        full_source: source,
-        rename_targets,
-    };
-
-    collect_script_content_edits(&edit_context, source, source_type, 0)
-}
-
-fn collect_script_content_edits(
-    context: &ScriptEditContext<'_>,
-    script_source: &str,
-    source_type: SourceType,
-    base_offset: usize,
-) -> Vec<TextEdit> {
-    let allocator = Allocator::default();
-    let parsed = Parser::new(&allocator, script_source, source_type).parse();
-
-    let mut collector = ModuleSpecifierCollector::default();
-    collector.visit_program(&parsed.program);
-
-    let Some(future_dir) = context.future_path.parent() else {
-        return Vec::new();
-    };
-
-    collector
-        .specifiers
-        .into_iter()
-        .filter_map(|specifier| {
-            let new_text = super::alias::rewrite_specifier(
-                context.state,
-                context.current_path,
-                future_dir,
-                &specifier.specifier,
-                context.rename_targets,
-            )?;
-
-            if new_text == specifier.specifier {
-                return None;
-            }
-
-            let start_offset = base_offset + specifier.start;
-            let end_offset = base_offset + specifier.end;
-            let range = offset_range(context.full_source, start_offset, end_offset)?;
-
-            Some(TextEdit { range, new_text })
-        })
-        .collect()
-}
-
-pub(super) fn rewrite_relative_specifier(
-    state: &ServerState,
-    current_importer_dir: &Path,
-    future_importer_dir: &Path,
-    specifier: &str,
-    rename_targets: &[RenameTarget],
-) -> Option<std::string::String> {
-    let (specifier_path, suffix) = split_specifier_suffix(specifier);
-    if !specifier_path.starts_with("./") && !specifier_path.starts_with("../") {
-        return None;
-    }
-
-    let mut selected = None;
-    for candidate in specifier_candidates(current_importer_dir, specifier_path) {
-        if candidate_exists(state, &candidate.resolved)
-            || apply_all_path_renames(&candidate.resolved, rename_targets).is_some()
-        {
-            selected = Some(candidate);
-            break;
-        }
-    }
-
-    let selected = selected?;
-    let future_target = apply_all_path_renames(&selected.resolved, rename_targets)
-        .unwrap_or_else(|| normalize_path_buf(&selected.resolved));
-
-    if future_target == normalize_path_buf(&selected.resolved)
-        && normalize_path_buf(current_importer_dir) == normalize_path_buf(future_importer_dir)
-    {
-        return None;
-    }
-
-    let mut rewritten =
-        render_module_specifier(future_importer_dir, &future_target, selected.style)?;
-    rewritten.push_str(suffix);
-    Some(rewritten)
-}
-
-fn render_module_specifier(
-    importer_dir: &Path,
-    target_path: &Path,
-    style: RenderStyle,
-) -> Option<std::string::String> {
-    let rendered_target = match style {
-        RenderStyle::Explicit => normalize_path_buf(target_path),
-        RenderStyle::Extensionless => strip_extension(target_path),
-        RenderStyle::DirectoryIndex => {
-            if is_index_file(target_path) {
-                normalize_path_buf(target_path.parent()?)
-            } else {
-                strip_extension(target_path)
-            }
-        }
-    };
-
-    relative_module_path(importer_dir, &rendered_target)
-}
-
-pub(super) fn relative_module_path(from_dir: &Path, to_path: &Path) -> Option<std::string::String> {
-    let from_dir = normalize_path_buf(from_dir);
-    let to_path = normalize_path_buf(to_path);
-
-    let from_components = from_dir.components().collect::<Vec<_>>();
-    let to_components = to_path.components().collect::<Vec<_>>();
-
-    let mut common = 0usize;
-    while common < from_components.len()
-        && common < to_components.len()
-        && from_components[common] == to_components[common]
-    {
-        common += 1;
-    }
-
-    if common == 0
-        && matches!(from_components.first(), Some(Component::Prefix(_)))
-        && matches!(to_components.first(), Some(Component::Prefix(_)))
-    {
-        return None;
-    }
-
-    let mut parts = Vec::new();
-    for _ in common..from_components.len() {
-        parts.push("..".to_string());
-    }
-    for component in &to_components[common..] {
-        let part = match component {
-            Component::Normal(value) => value.to_string_lossy().to_string(),
-            Component::CurDir => ".".to_string(),
-            Component::ParentDir => "..".to_string(),
-            Component::RootDir | Component::Prefix(_) => continue,
-        };
-        parts.push(part);
-    }
-
-    let joined = if parts.is_empty() {
-        ".".to_string()
-    } else {
-        parts.join("/")
-    };
-
-    if joined.starts_with("../") || joined == ".." {
-        Some(joined)
-    } else if joined == "." {
-        Some("./".to_string())
-    } else if joined.starts_with("./") {
-        Some(joined)
-    } else {
-        let mut prefixed = std::string::String::from("./");
-        prefixed.push_str(&joined);
-        Some(prefixed)
-    }
-}
-
-fn specifier_candidates(importer_dir: &Path, specifier: &str) -> Vec<SpecifierCandidate> {
-    let resolved = normalize_path_buf(&importer_dir.join(specifier));
-    let mut candidates = Vec::new();
-
-    if Path::new(specifier).extension().is_none() {
-        for extension in RESOLVABLE_SCRIPT_EXTENSIONS {
-            candidates.push(SpecifierCandidate {
-                resolved: resolved.with_extension(extension),
-                style: RenderStyle::Extensionless,
-            });
-        }
-
-        for extension in RESOLVABLE_SCRIPT_EXTENSIONS {
-            let mut index_name = std::string::String::from("index.");
-            index_name.push_str(extension);
-            candidates.push(SpecifierCandidate {
-                resolved: resolved.join(index_name),
-                style: RenderStyle::DirectoryIndex,
-            });
-        }
-
-        candidates.push(SpecifierCandidate {
-            resolved,
-            style: RenderStyle::Explicit,
-        });
-    } else {
-        candidates.push(SpecifierCandidate {
-            resolved,
-            style: RenderStyle::Explicit,
-        });
-    }
-
-    candidates
-}
-
 fn rename_targets(renames: &[FileRename]) -> Vec<RenameTarget> {
     renames
         .iter()
@@ -618,46 +238,6 @@ fn apply_all_uri_renames(uri: &Url, renames: &[RenameTarget]) -> Option<Url> {
     let path = uri.to_file_path().ok()?;
     let path = apply_all_path_renames(&path, renames)?;
     Url::from_file_path(path).ok()
-}
-
-pub(super) fn apply_all_path_renames(path: &Path, renames: &[RenameTarget]) -> Option<PathBuf> {
-    let mut updated = normalize_path_buf(path);
-    let mut changed = false;
-
-    for rename in renames {
-        if let Some(next) = apply_path_rename(&updated, rename) {
-            updated = next;
-            changed = true;
-        }
-    }
-
-    if changed { Some(updated) } else { None }
-}
-
-fn apply_path_rename(path: &Path, rename: &RenameTarget) -> Option<PathBuf> {
-    let normalized = normalize_path_buf(path);
-    if normalized == rename.old_path {
-        return Some(rename.new_path.clone());
-    }
-
-    let suffix = normalized.strip_prefix(&rename.old_path).ok()?;
-    if suffix.as_os_str().is_empty() {
-        Some(rename.new_path.clone())
-    } else {
-        Some(normalize_path_buf(&rename.new_path.join(suffix)))
-    }
-}
-
-pub(super) fn candidate_exists(state: &ServerState, path: &Path) -> bool {
-    if path.exists() {
-        return true;
-    }
-
-    let Ok(uri) = Url::from_file_path(path) else {
-        return false;
-    };
-
-    state.documents.contains(&uri)
 }
 
 fn importer_kind(path: &Path, only_vue_importers: bool) -> Option<ImporterKind> {
@@ -694,11 +274,6 @@ fn snapshot_open_documents(state: &ServerState) -> OpenDocumentSnapshots {
     snapshots
 }
 
-pub(super) fn split_specifier_suffix(specifier: &str) -> (&str, &str) {
-    let split_at = specifier.find(['?', '#']).unwrap_or(specifier.len());
-    (&specifier[..split_at], &specifier[split_at..])
-}
-
 pub(super) fn offset_range(source: &str, start: usize, end: usize) -> Option<Range> {
     let (start_line, start_character) = offset_to_position(source, start);
     let (end_line, end_character) = offset_to_position(source, end);
@@ -713,49 +288,6 @@ pub(super) fn offset_range(source: &str, start: usize, end: usize) -> Option<Ran
             character: end_character,
         },
     })
-}
-
-fn script_source_type(lang: Option<&str>) -> SourceType {
-    match lang.unwrap_or("js") {
-        "ts" => SourceType::ts(),
-        "tsx" => SourceType::tsx(),
-        "jsx" => SourceType::jsx(),
-        _ => SourceType::mjs(),
-    }
-}
-
-pub(super) fn normalize_path_buf(path: &Path) -> PathBuf {
-    let mut normalized = PathBuf::new();
-
-    for component in path.components() {
-        match component {
-            Component::CurDir => {}
-            Component::ParentDir => {
-                normalized.pop();
-            }
-            Component::RootDir | Component::Prefix(_) | Component::Normal(_) => {
-                normalized.push(component.as_os_str());
-            }
-        }
-    }
-
-    normalized
-}
-
-pub(super) fn strip_extension(path: &Path) -> PathBuf {
-    let mut stripped = normalize_path_buf(path);
-    let path = stripped.to_string_lossy();
-    let extension_depth = 1 + usize::from(
-        path.ends_with(".d.ts") || path.ends_with(".d.mts") || path.ends_with(".d.cts"),
-    );
-    for _ in 0..extension_depth {
-        let _ = stripped.set_extension("");
-    }
-    stripped
-}
-
-fn is_index_file(path: &Path) -> bool {
-    strip_extension(path).ends_with("index")
 }
 
 #[cfg(feature = "native")]
